physical/hpos: infer GRUB device
See 4223456cde98b669199ee35345641a0e7163d774.

Instead of using /dev/sdX as GRUB device, we are now inferring GRUB
device based on which device root partition is located at. In both
HoloPort and HoloPort+, rootfs partition is located on GRUB boot device.

We have to go with this solution because there is no other way to tell
where bootloader is supposed to be installed that would be compatible
with our factory image.